requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
requests==2.32.3
beautifulsoup4==4.12.3
lxml==5.3.0
python-multipart==0.0.9
duckduckgo-search>=4.4.0
python-telegram-bot==21.10
python-dotenv==1.0.1
curl-cffi==0.9.0
google-genai==0.3.0